Refactor to move out of categorical column values script
