+++ utils/validation/validation_rules/locations_api_raw_validation_rules.py
@@ -3,6 +3,9 @@ from dataclasses import dataclass
 from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns as CQCL,
 )
+from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+    CqcLocationCleanedColumns as CQCLClean,
+)
 from utils.column_names.ind_cqc_pipeline_columns import (
     PartitionKeys as Keys,
 )
@@ -12,6 +15,15 @@ from utils.validation.categorical_column_values import (
     CQCCategoricalValues,
     CQCDistinctValues,
 )
+from utils.column_values.categorical_column_values import (
+    Dormancy,
+    RegistrationStatus,
+    PrimaryServiceType,
+)
+
+dormancy_column_values = Dormancy(CQCL.dormancy, contains_null_values=True)
+registration_status_column_values = RegistrationStatus(CQCL.registration_status)
+primary_service_type_column_values = PrimaryServiceType(CQCLClean.primary_service_type)
 
 
 @dataclass
@@ -38,12 +50,12 @@ class LocationsAPIRawValidationRules:
         },
         RuleName.categorical_values_in_columns: {
             CQCL.care_home: CQCCategoricalValues.care_home_values,
-            CQCL.registration_status: CQCCategoricalValues.registration_status_raw,
-            CQCL.dormancy: CQCCategoricalValues.dormancy_values,
+            CQCL.registration_status: registration_status_column_values.categorical_values,
+            CQCL.dormancy: dormancy_column_values.categorical_values,
         },
         RuleName.distinct_values: {
             CQCL.care_home: CQCDistinctValues.care_home_values,
-            CQCL.registration_status: CQCDistinctValues.registration_status_raw_values,
-            CQCL.dormancy: CQCDistinctValues.dormancy_values,
+            CQCL.registration_status: registration_status_column_values.count_of_categorical_values,
+            CQCL.dormancy: dormancy_column_values.count_of_categorical_values,
         },
     }
